Added notion of edges and tree. Major refactor
